feat: cottage and boat owner registration, CRUD for cottages, cottages search, cottage reservation history view, client view, viewing and updating of personal info, password changing and account deleting for cottage owner

+++ Backend/ftn/src/main/java/com/fishingbooker/ftn/service/CottageServiceImpl.java
@@ -1,16 +1,30 @@
 package com.fishingbooker.ftn.service;
 
+import com.fishingbooker.ftn.bom.Address;
+import com.fishingbooker.ftn.bom.RuleOfConduct;
+import com.fishingbooker.ftn.bom.adventures.AdventureUtility;
 import com.fishingbooker.ftn.bom.cottages.Cottage;
+import com.fishingbooker.ftn.bom.cottages.CottageUtility;
 import com.fishingbooker.ftn.bom.cottages.Room;
+import com.fishingbooker.ftn.bom.users.CottageOwner;
 import com.fishingbooker.ftn.conversion.DataConverter;
+import com.fishingbooker.ftn.dto.CottageCreationDto;
 import com.fishingbooker.ftn.dto.CottageDto;
+import com.fishingbooker.ftn.dto.RoomDto;
 import com.fishingbooker.ftn.repository.*;
 import com.fishingbooker.ftn.service.interfaces.CottageService;
+import com.fishingbooker.ftn.service.interfaces.ImageService;
+import com.fishingbooker.ftn.service.interfaces.RuleOfConductService;
+import com.fishingbooker.ftn.service.interfaces.UtilityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.io.IOException;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -23,6 +37,9 @@ public class CottageServiceImpl implements CottageService {
     private final CottageOwnerRepository cottageOwnerRepository;
     private final CottageUtilityRepository utilityRepository;
     private final DataConverter converter;
+    private final RuleOfConductService ruleOfConductService;
+    private final ImageService imageService;
+    private final UtilityService utilityService;
 
     @Override
     public List<Cottage> findAll() {
@@ -36,9 +53,73 @@ public class CottageServiceImpl implements CottageService {
     }
 
     @Override
-    public Long delete(Long id) {
-        cottageRepository.deleteById(id);
-        return id;
+    public Cottage get(Long id) {
+        return cottageRepository.get(id);
     }
 
+    @Override
+    public List<CottageDto> findByCottageOwnerId(long id) {
+        return converter.convert(cottageRepository.findByCottageOwnerId(id), CottageDto.class);
+    }
+
+    @Override
+    public Long create(CottageCreationDto cottageDto) {
+        Cottage cottage;
+        if (cottageDto.getId()==-1){
+            cottage=new Cottage();
+        }else{
+            cottage=cottageRepository.getById(cottageDto.getId());
+        }
+
+        CottageOwner cottageOwner=cottageOwnerRepository.getById(cottageDto.getOwnerId());
+        Address address=new Address();
+        cottage.setName(cottageDto.getName());
+        address.setCity(cottageDto.getCity());
+        address.setCountry(cottageDto.getCountry());
+        address.setStreet(cottageDto.getStreet());
+        cottage.setAddress(address);
+        cottage.setPrice(cottageDto.getPrice());
+        cottage.setCottageOwner(cottageOwner);
+        List<RuleOfConduct> rules=ruleOfConductService.createRulesFromString(cottageDto.getRules());
+        cottage.setRules(new HashSet<>(rules));
+        cottage.setDescription(cottageDto.getDescription());
+        cottage.setGuestLimit(cottageDto.getGuestLimit());
+        try {
+            cottage.setImages(imageService.saveImages(cottageDto.getImages()));
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        Cottage savedCottage=cottageRepository.save(cottage);
+        savedCottage.setRooms(convertDtoToModel(cottageDto.getRooms(), savedCottage));
+        Set<CottageUtility> utilities=utilityService.convertStringToUtility(cottageDto.getAdditionalServices(), savedCottage);
+        savedCottage.setUtilities(utilities);
+        return cottageRepository.save(savedCottage).getId();
+    }
+
+    private Set<Room> convertDtoToModel(List<RoomDto> rooms, Cottage cottage) {
+        return rooms.stream().map(roomDto -> createRoom(roomDto, cottage)).collect(Collectors.toSet());
+    }
+
+    private Room createRoom(RoomDto roomDto, Cottage cottage) {
+        if (roomDto.getId()==-1){
+            return new Room(roomDto.getNumberOfBeds(), cottage);
+        }else{
+            Room room=new Room();
+            room.setId(roomDto.getId());
+            room.setNumberOfBeds(roomDto.getNumberOfBeds());
+            room.setCottage(cottage);
+            return room;
+        }
+    }
+
+    @Override
+    public Boolean delete(Long id) {
+        Cottage cottage = cottageRepository.getById(id);
+        if ((cottage.getCottageReservations()==null || cottage.getCottageReservations().size()==0)
+                && (cottage.getCottageQuickReservations()==null || cottage.getCottageQuickReservations().size()==0)) {
+            cottageRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
 }
